+++ NexusKrop.IceCube/Data/KeyValueContainer.io.cs
@@ -15,6 +15,7 @@
 namespace NexusKrop.IceCube.Data;
 
 using NexusKrop.IceCube.IO;
+using NexusKrop.IceCube.Util.Enumerables;
 using System;
 using System.Diagnostics;
 using System.IO;